feat: melhora a componentização da página de busca
